refactor(cache): keep async filtering state from sliding its TTL and move newStmtCache beside its SQLite caller
